requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
dataclasses-json==0.5.14
decorator==5.1.1
dnspython==2.4.2
docopt==0.6.2
effdet==0.4.1
emoji==2.8.0
faiss-cpu==1.7.4
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
grpcio==1.59.0
grpcio-tools==1.59.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.16.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.310
langdetect==1.0.9
langsmith==0.0.43
layoutparser==0.3.4
llama-index==0.8.40
loguru==0.5.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.2
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.16.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.8.1.78
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==9.5.0
pinecone-client @ git+https://github.com/pinecone-io/pinecone-python-client.git@212299da0188827b57bfe7761f5dc4c2046446c0
pipreqs==0.4.13
plotly==5.17.0
portalocker==2.8.2
protobuf==4.23.4
pyarrow==12.0.1
PyAudio==0.2.13
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.1.1
pypdf==3.16.2
PyPDF2==3.0.1
pypdfium2==4.20.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
qdrant-client==1.5.4
rapidfuzz==3.3.1
referencing==0.30.1
regex==2023.6.3
reqs==0.0.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.19
streamlit==1.27.2
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.1
timm==0.9.7
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
twilio==8.9.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.10.19
unstructured-inference==0.6.6
unstructured.pytesseract==0.3.12
urllib3==1.26.17
validators==0.20.0
watchdog==3.0.0
win32-setctime==1.1.0
yarg==0.1.9
yarl==1.9.2
zipp==3.16.2
